LayoutTests/accessibility/mac/aria-expanded-standard-items-expected.txt

foo
bar
This tests that ARIA expanded will be exposed on arbitrary elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.role is 'AXRole: AXButton'
PASS accessibilityController.focusedElement.isExpanded is true
PASS accessibilityController.focusedElement.isExpanded is true
PASS accessibilityController.focusedElement.childAtIndex(0).isExpanded is false
PASS accessibilityController.focusedElement.childAtIndex(0).role is 'AXRole: AXStaticText'
PASS successfullyParsed is true

TEST COMPLETE
